Fix disproportionality sort CCG-719; Fix retina size for 100k chart
